refactor (chat): hook state on chat (NOTIF-01)
